feat: integrate ioredis for redis sentinel support
