+++ apps/api/README.md
@@ -1,0 +1,22 @@
+## API
+
+Standalone server which serves the indexed data
+
+### Config
+
+Update environment variables in `*.api.env`
+
+```
+DATABASE_URL=
+USER_DB_URL=
+REDIS_SENTINEL=
+RPC_URL=
+NETWORK=
+
+# Optional
+DATABASE_CA=
+DATABASE_CERT=
+DATABASE_KEY=
+REDIS_URL=
+SENTRY_DSN=
+```
