+++ apps/api/src/index.ts
@@ -1,6 +1,6 @@
 import config from '#config';
 import logger from '#libs/logger';
-import redis from '#libs/redis';
+import redis, { userRedisClient } from '#libs/redis';
 import sentry from '#libs/sentry';
 
 import app from './app.js';
@@ -9,15 +9,18 @@ app.listen(config.port, async () => {
   logger.info(`server listening on port ${config.port}`);
 });
 
-const onSignal = (signal: number | string) => {
-  const handler = async () => {
-    await redis.quit();
-    await sentry.close(1000);
-  };
+const onSignal = async (signal: number | string) => {
+  try {
+    await Promise.all([
+      redis.quit(),
+      userRedisClient.quit(),
+      sentry.close(1000),
+    ]);
+  } catch (error) {
+    logger.error(error);
+  }
 
-  handler()
-    .catch(logger.error)
-    .finally(() => process.kill(process.pid, signal));
+  process.kill(process.pid, signal);
 };
 
 process.once('SIGINT', onSignal);
