+++ apps/api/src/services/txns.ts
@@ -2,7 +2,7 @@ import { Response } from 'express';
 
 import catchAsync from '#libs/async';
 import db from '#libs/db';
-import { cache } from '#libs/redis';
+import redis from '#libs/redis';
 import { Count, Item, Latest, List } from '#libs/schema/txns';
 import { getPagination, keyBinder } from '#libs/utils';
 import { RequestValidator } from '#types/types';
@@ -246,7 +246,7 @@ const latest = catchAsync(
       { limit },
     );
 
-    const txns = await cache(
+    const txns = await redis.cache(
       'txns:latest',
       async () => {
         try {
@@ -257,7 +257,7 @@ const latest = catchAsync(
           return null;
         }
       },
-      { EX: EXPIRY },
+      EXPIRY,
     );
 
     return res.status(200).json({ txns });
