+++ apps/backend/src/config.ts
@@ -1,4 +1,4 @@
-import { cleanEnv, str } from 'envalid';
+import { cleanEnv, str, url } from 'envalid';
 
 import { Network } from 'nb-types';
 
@@ -15,7 +15,8 @@ const env = cleanEnv(process.env, {
   NETWORK: str({
     choices: [Network.MAINNET, Network.TESTNET],
   }),
-  REDIS_URL: str(),
+  REDIS_SENTINEL: str({ default: '' }),
+  REDIS_URL: url({ default: '' }),
   RPC_URL: str(),
   RPC_URL2: str(),
   SENTRY_DSN: str({ default: '' }),
@@ -37,6 +38,7 @@ const config: Config = {
   genesisHeight,
   lcwApiKey: env.LIVECOINWATCH_API_KEY,
   network: env.NETWORK,
+  redisSentinel: env.REDIS_SENTINEL,
   redisUrl: env.REDIS_URL,
   rpcUrl: env.RPC_URL,
   rpcUrl2: env.RPC_URL2,
