+++ apps/indexer-balance/src/types/types.ts
@@ -12,7 +12,6 @@ export type Config = {
   insertLimit: number;
   network: Network;
   preloadSize: number;
-  redisUrl: string;
   rpcUrl: string;
   s3BucketName: string;
   s3Endpoint: null | types.EndpointConfig;
