+++ apps/indexer-base/src/config.ts
@@ -14,7 +14,8 @@ const env = cleanEnv(process.env, {
   NETWORK: str({
     choices: [Network.MAINNET, Network.TESTNET],
   }),
-  REDIS_URL: str(),
+  REDIS_SENTINEL: str({ default: '' }),
+  REDIS_URL: url({ default: '' }),
   S3_ENDPOINT: url({ default: '' }),
   SENTRY_DSN: str({ default: '' }),
 });
@@ -54,6 +55,7 @@ const config: Config = {
   insertLimit: 1_000, // records to insert into the db at a time
   network: env.NETWORK,
   preloadSize: 100, // blocks to preload in nearlake
+  redisSentinel: env.REDIS_SENTINEL,
   redisUrl: env.REDIS_URL,
   s3BucketName,
   s3Endpoint,
