+++ apps/indexer-base/src/index.ts
@@ -2,7 +2,6 @@ import { logger } from 'nb-logger';
 
 import config from '#config';
 import knex from '#libs/knex';
-import { redisClient } from '#libs/redis';
 import sentry from '#libs/sentry';
 import { syncGenesis } from '#services/genesis';
 import { syncData } from '#services/stream';
@@ -10,7 +9,6 @@ import { syncData } from '#services/stream';
 (async () => {
   try {
     logger.info({ network: config.network }, 'initializing base indexer...');
-    await redisClient.connect();
     logger.info('syncing genesis data...');
     await syncGenesis();
     logger.info('syncing blockchain data...');
@@ -25,9 +23,7 @@ import { syncData } from '#services/stream';
 
 const onSignal = async (signal: number | string) => {
   try {
-    await knex.destroy();
-    await redisClient.quit();
-    await sentry.close(1_000);
+    await Promise.all([knex.destroy(), sentry.close(1_000)]);
   } catch (error) {
     logger.error(error);
   }
