+++ apps/indexer-events/package.json
@@ -22,8 +22,7 @@
     "envalid": "8.0.0",
     "globby": "13.2.2",
     "lodash-es": "4.17.21",
-    "near-lake-framework": "2.0.0",
-    "redis-lock": "1.0.0"
+    "near-lake-framework": "2.0.0"
   },
   "devDependencies": {
     "@types/lodash-es": "~4.17",
@@ -33,7 +32,6 @@
     "nb-knex": "*",
     "nb-logger": "*",
     "nb-near": "*",
-    "nb-redis": "*",
     "nb-tsconfig": "*",
     "nb-types": "*",
     "nb-utils": "*",
