--- apps/indexer-events/src/libs/redis.ts
@@ -1,11 +1,0 @@
-import redisLock from 'redis-lock';
-
-import { Redis } from 'nb-redis';
-
-import config from '#config';
-
-export const redis = new Redis(`events:${config.network}`, {
-  url: config.redisUrl,
-});
-export const redisClient = redis.getClient();
-export const redLock = redisLock(redisClient);
