--- apps/indexer-events/src/types/redis-lock.d.ts
@@ -1,20 +1,0 @@
-declare module 'redis-lock' {
-  import {
-    RedisClientType,
-    RedisDefaultModules,
-    RedisFunctions,
-    RedisModules,
-    RedisScripts,
-  } from 'redis';
-
-  function redisLock(
-    client: RedisClientType<
-      RedisDefaultModules & RedisModules,
-      RedisFunctions,
-      RedisScripts
-    >,
-    retryDelay?: number,
-  ): (lockName: string, timeout?: number) => Promise<() => Promise<void>>;
-
-  export default redisLock;
-}
