+++ mainnet.api.env.example
@@ -1,7 +1,9 @@
 DATABASE_URL=
-REDIS_URL=
 RPC_URL=
 NETWORK=mainnet
+
+REDIS_SENTINEL="localhost:26379|localhost:26380"
+REDIS_URL="localhost:6379"
 
 USER_DB_URL=
 USER_REDIS_URL=
