+++ mainnet.backend.env.example
@@ -3,6 +3,9 @@ RPC_URL=
 RPC_URL2=
 NETWORK=mainnet
 
+REDIS_SENTINEL="localhost:26379|localhost:26380"
+REDIS_URL="localhost:6379"
+
 DATABASE_CA=
 DATABASE_CERT=
 DATABASE_KEY=
