+++ packages/nb-redis/package.json
@@ -11,7 +11,7 @@
     "lint": "tsc --noEmit && eslint ./ --fix"
   },
   "dependencies": {
-    "redis": "4.6.10"
+    "ioredis": "5.3.2"
   },
   "devDependencies": {
     "eslint-config-custom-node": "*",
